+++ lib/loading_indicator.dart
@@ -1,3 +1,9 @@
 library loading_indicator;
 
+export 'package:loading_indicator/src/control/loading_indicator_controller.dart'
+    show
+        LoadingIndicatorCommandCanceled,
+        LoadingIndicatorController,
+        LoadingIndicatorPauseBehavior,
+        LoadingIndicatorPlaybackStatus;
 export 'package:loading_indicator/src/loading.dart';