+++ lib/src/indicators/ball_clip_rotate_multiple.dart
@@ -1,7 +1,7 @@
 import 'dart:math';
 
 import 'package:flutter/material.dart';
-import 'package:loading_indicator/src/indicators/base/indicator_controller.dart';
+import 'package:loading_indicator/src/control/loading_indicator_controller.dart';
 import 'package:loading_indicator/src/shape/indicator_painter.dart';
 import 'package:loading_indicator/src/transition/matrix4_transform.dart';