+++ lib/src/indicators/line_scale_pulse_out_rapid.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
+import 'package:loading_indicator/src/control/loading_indicator_controller.dart';
 import 'package:loading_indicator/src/decorate/decorate.dart';
-import 'package:loading_indicator/src/indicators/base/indicator_controller.dart';
 import 'package:loading_indicator/src/shape/indicator_painter.dart';
 
 /// LineScalePulseOutRapid.